Relocate code. Increase readbility
